Converts a AnnData object to a MAMS object

Converts a AnnData object to a MAMS object

Usage

convert_AnnData_to_MAMS(
  object_list,
  X_processing,
  observation_subsets,
  dataset_id
)

Arguments

object_list

A named list of AnnData objects objects to be converted to MAMS format

X_processing

A vector of the same length as object_list, where each element describes the processing for X in each object. For example, "counts" or "logcounts" or "scaled".

observation_subsets

A vector with same length as object_list indicating the observation subset name for each AnnData object. One of: full, filtered, threshold, detected, nonartifact, clean.

dataset_id

Specify id of the dataset

Value

A MAMS object containing all the extracted metadata fields.

Examples

if (FALSE) {
library(reticulate)
anndata <- import("anndata")
counts <- matrix(rpois((500*200), 1), nrow = 500, ncol = 200, 
    dimnames = list(paste0("Row", 1:500), paste0("Col", 1:200)))
adata <- anndata$AnnData(X = counts)
py_set_item(adata$obsm, 'X_pca', matrix(rnorm(50), nrow=500, ncol=10))
py_set_item(adata$obsm, 'X_tsne', matrix(rnorm(50), nrow=500, ncol=2))
mams <- convert_AnnData_to_MAMS(object_list = list(adata = adata), 
    observation_subsets = c("full"), dataset_id = "dataset1")
print(mams)
}